+++ R/zzz.R
@@ -1,5 +1,48 @@
 .onLoad <- function (libname, pkgname) {
   
+  packageStartupMessage(sprintf("polmineR v%s", packageVersion("polmineR")))
+
+  # The package includes the files configure / configure.win in the top-level
+  # directory that will set paths in the registry files for corpora correctly
+  # upon installation. However, configure / configure.win may not be executed
+  # when a pre-compiled package is distributed. Therefore, the following lines
+  # check upon whenever loading polmineR whether the paths are set correctly,
+  # and adjust them if necessary.
+  pkg_cwb_dir <- file.path(libname, pkgname, "extdata", "cwb")
+  pkg_registry_dir <- file.path(pkg_cwb_dir, "registry")
+  pkg_indexed_corpora_dir <- file.path(pkg_cwb_dir, "indexed_corpora")
+  
+  for (corpus in list.files(pkg_registry_dir)){
+    registry_file <- file.path(pkg_registry_dir, corpus)
+    registry <- readLines(registry_file)
+    
+    home_line_no <- grep("^HOME", registry)
+    info_line_no <- grep("^INFO", registry)
+    registry_home_dir <- gsub('^HOME\\s+"*(.*?)"*\\s*$', "\\1", registry[home_line_no])
+    registry_info_file <- gsub('^INFO\\s+"*(.*?)"*\\s*$', "\\1", registry[info_line_no])
+    
+    pkg_home_dir <- file.path(pkg_indexed_corpora_dir, corpus)
+    if (!identical(x = registry_home_dir, y = pkg_home_dir)){
+      packageStartupMessage(sprintf("data directory set for: %s", corpus))
+      info_file_new <- file.path(pkg_home_dir, basename(registry_info_file), fsep = "/")
+      # quote paths on Windows, and unix-likes, if path includes whitespace
+      if (.Platform$OS.type == "windows"){
+        registry[home_line_no] <- sprintf('HOME "%s"', pkg_home_dir)
+        registry[info_line_no] <- sprintf('INFO "%s"', info_file_new)
+      } else {
+        if (grepl("\\s+", pkg_home_dir)){
+          registry[grep("^HOME", registry)] <- sprintf('HOME "%s"', pkg_home_dir)
+          registry[info_line_no] <- sprintf('INFO "%s"', info_file_new)
+        } else {
+          registry[grep("^HOME", registry)] <- sprintf("HOME %s", pkg_home_dir)
+          registry[info_line_no] <- sprintf("INFO %s", info_file_new)
+        }
+      }
+      writeLines(text = registry, con = registry_file, sep = "\n")
+    }
+  }
+
+  
   # polmineR:::CQI - assign it to package namespace
   CQI <- switch(
     Sys.getenv("POLMINER_INTERFACE"),
@@ -12,13 +55,12 @@
   # if environment variable CORPUS_REGISTRY is not set, use data in the polmineR package
   # this needs to be done after assigning CQI, as registry_reset will call setTemplate
   if (Sys.getenv("CORPUS_REGISTRY") %in% c("", "/")){
-    polmineRPackageRegistry <- file.path(libname, pkgname, "extdata", "cwb", "registry")
-    Sys.setenv("CORPUS_REGISTRY" = polmineRPackageRegistry)
-    registry_reset(registryDir = polmineRPackageRegistry, verbose = FALSE)
+    Sys.setenv("CORPUS_REGISTRY" = pkg_registry_dir)
+    registry_reset(registryDir = pkg_registry_dir, verbose = FALSE)
   }
   
   options(
-    "polmineR.corpus" = "GERMAPARLMINI",
+    "polmineR.corpus" = "REUTERS",
     "polmineR.pAttribute" = "word",
     "polmineR.left" = 5,
     "polmineR.right" = 5,
@@ -47,7 +89,6 @@
   # initializing CQP by calling RcppCWB::cqp_initialize would logically done here,
   # but for some (unknown) reason, a package crash occurrs, when CQP is initialized
   # on attach - thus, the 'on demand'-solution (call cqp_initialize before calling cqp_query)
-  packageStartupMessage(sprintf("polmineR %s", packageVersion("polmineR")))
   packageStartupMessage("registry:  ", getOption("polmineR.defaultRegistry"))
   packageStartupMessage("interface: ", if (exists("CQI")) class(CQI)[1] else "not set")
 }